modify: ReservationState 위치 이동
- App -> Hero로 옮김
- 불필요한 리렌더가 많은 것 같고, 여기에 있어도 괜찮을 것 같아서 수정
