+++ fe/src/App.tsx
@@ -7,39 +7,39 @@ import { ReservationStateContext, ReservationDispatchContext } from './Contexts'
 import { ThemeProvider } from 'styled-components';
 import { theme } from './styles/theme';
 
-const initialState = {
-    location: {
-        id: 0,
-        city: '',
-    },
-    checkIn: {
-        year: 0,
-        month: 0,
-        day: 0,
-    },
-    checkOut: {
-        year: 0,
-        month: 0,
-        day: 0,
-    },
-    fee: 0,
-    people: {
-        guest: 0,
-        kids: 0,
-    },
-} as ReservationContext;
+// const initialState = {
+//     location: {
+//         id: 0,
+//         city: '',
+//     },
+//     checkIn: {
+//         year: 0,
+//         month: 0,
+//         day: 0,
+//     },
+//     checkOut: {
+//         year: 0,
+//         month: 0,
+//         day: 0,
+//     },
+//     fee: 0,
+//     people: {
+//         guest: 0,
+//         kids: 0,
+//     },
+// } as ReservationContext;
 
 function App(): React.ReactElement {
-    const [reservationState, reservationDispatch] = useReducer(reservationReducer, initialState);
+    // const [reservationState, reservationDispatch] = useReducer(reservationReducer, initialState);
 
     return (
-        <ReservationDispatchContext.Provider value={reservationDispatch}>
-            <ReservationStateContext.Provider value={reservationState}>
-                <ThemeProvider theme={theme}>
-                    <Router />
-                </ThemeProvider>
-            </ReservationStateContext.Provider>
-        </ReservationDispatchContext.Provider>
+        // <ReservationDispatchContext.Provider value={reservationDispatch}>
+        //     <ReservationStateContext.Provider value={reservationState}>
+        <ThemeProvider theme={theme}>
+            <Router />
+        </ThemeProvider>
+        //     </ReservationStateContext.Provider>
+        // </ReservationDispatchContext.Provider>
     );
 }
